Added dynamic pages for server and user search. Need to hookup api search to load in data
