+++ next.config.ts
@@ -3,7 +3,7 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
   /* config options here */
   // basePath: "/LeagueWorstPerformanceTracker", FOR GIT PAGES ONLY
-  output: "export",  // <=== enables static exports
+  // output: "export",  // <=== enables static exports
   reactStrictMode: true,
 };
