+++ src/app/page.tsx
@@ -1,9 +1,9 @@
+import SearchBar from "@/components/SearchBar";
 import SummonerInfo from "@/components/SummonerInfo";
 
 export default function Home() {
   return (
-    <div className="flex justify-self-center w-2/3">
-      <SummonerInfo />
+    <div className="flex justify-self-center w-full md:w-2/3">
     </div>
   );
 }
